+++ tests/Hydra/Serializer/PartialCollectionViewNormalizerTest.php
@@ -17,6 +17,8 @@ use ApiPlatform\Core\DataProvider\PaginatorInterface;
 use ApiPlatform\Core\DataProvider\PartialPaginatorInterface;
 use ApiPlatform\Core\Hydra\Serializer\PartialCollectionViewNormalizer;
 use ApiPlatform\Core\Metadata\Resource\Factory\ResourceMetadataFactoryInterface;
+use ApiPlatform\Core\Metadata\Resource\ResourceMetadata;
+use ApiPlatform\Core\Tests\Fixtures\TestBundle\Entity\SoMany;
 use ApiPlatform\Core\Tests\ProphecyTrait;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
@@ -79,13 +81,31 @@ class PartialCollectionViewNormalizerTest extends TestCase
                     '@id' => '/?_page=3',
                     '@type' => 'hydra:PartialCollectionView',
                     'hydra:previous' => '/?_page=2',
+                    'hydra:next' => '/?_page=4',
                 ],
             ],
             $this->normalizePaginator(true)
         );
     }
 
-    private function normalizePaginator($partial = false)
+    public function testNormalizeWithCursorBasedPagination(): void
+    {
+        self::assertEquals(
+            [
+                'foo' => 'bar',
+                'hydra:totalItems' => 40,
+                'hydra:view' => [
+                    '@id' => '/',
+                    '@type' => 'hydra:PartialCollectionView',
+                    'hydra:previous' => '/?id%5Bgt%5D=1',
+                    'hydra:next' => '/?id%5Blt%5D=2',
+                ],
+            ],
+            $this->normalizePaginator(false, true)
+        );
+    }
+
+    private function normalizePaginator(bool $partial = false, bool $cursor = false)
     {
         $paginatorProphecy = $this->prophesize($partial ? PartialPaginatorInterface::class : PaginatorInterface::class);
         $paginatorProphecy->getCurrentPage()->willReturn(3)->shouldBeCalled();
@@ -102,11 +122,33 @@ class PartialCollectionViewNormalizerTest extends TestCase
 
         $decoratedNormalizerProphecy = $this->prophesize(NormalizerInterface::class);
         $decoratedNormalizerProphecy->normalize(Argument::type($partial ? PartialPaginatorInterface::class : PaginatorInterface::class), null, Argument::type('array'))->willReturn($decoratedNormalize)->shouldBeCalled();
-        $resourceMetadataFactory = $this->prophesize(ResourceMetadataFactoryInterface::class);
 
-        $normalizer = new PartialCollectionViewNormalizer($decoratedNormalizerProphecy->reveal(), '_page', 'pagination', $resourceMetadataFactory->reveal());
+        $resourceMetadataFactoryProphecy = null;
 
-        return $normalizer->normalize($paginatorProphecy->reveal());
+        if ($cursor) {
+            $firstSoMany = new SoMany();
+            $firstSoMany->id = 1;
+            $firstSoMany->content = 'SoMany #1';
+
+            $lastSoMany = new SoMany();
+            $lastSoMany->id = 2;
+            $lastSoMany->content = 'SoMany #2';
+
+            $paginatorProphecy->rewind()->willReturn()->shouldBeCalledOnce();
+            $paginatorProphecy->valid()->willReturn(true, true, false)->shouldBeCalledTimes(3);
+            $paginatorProphecy->key()->willReturn(1, 2)->shouldBeCalledTimes(2);
+            $paginatorProphecy->current()->willReturn($firstSoMany, $lastSoMany)->shouldBeCalledTimes(2);
+            $paginatorProphecy->next()->willReturn()->shouldBeCalledTimes(2);
+
+            $soManyMetadata = new ResourceMetadata(null, null, null, null, ['get' => ['pagination_via_cursor' => [['field' => 'id', 'direction' => 'desc']]]]);
+
+            $resourceMetadataFactoryProphecy = $this->prophesize(ResourceMetadataFactoryInterface::class);
+            $resourceMetadataFactoryProphecy->create(SoMany::class)->willReturn($soManyMetadata)->shouldBeCalledOnce();
+        }
+
+        $normalizer = new PartialCollectionViewNormalizer($decoratedNormalizerProphecy->reveal(), '_page', 'pagination', $resourceMetadataFactoryProphecy ? $resourceMetadataFactoryProphecy->reveal() : null);
+
+        return $normalizer->normalize($paginatorProphecy->reveal(), null, ['resource_class' => SoMany::class, 'collection_operation_name' => 'get']);
     }
 
     public function testSupportsNormalization()